Change the log processing policy.

+++ delicate-scheduler/src/db/common/state.rs
@@ -7,15 +7,15 @@ pub mod task {
 }
 
 pub mod task_log {
-    pub const RUNNING: usize = 1 << 0;
+    pub const RUNNING: i16 = 1 << 0;
 
-    pub const NORMAL_ENDING: usize = 1 << 1;
+    pub const NORMAL_ENDING: i16 = 1 << 1;
 
-    pub const ABNORMAL_ENDING: usize = 1 << 2;
+    pub const ABNORMAL_ENDING: i16 = 1 << 2;
 
-    pub const TIMEOUT_ENDING: usize = 1 << 3;
+    pub const TIMEOUT_ENDING: i16 = 1 << 3;
 
-    pub const TMANUAL_CANCELLATION: usize = 1 << 4;
+    pub const TMANUAL_CANCELLATION: i16 = 1 << 4;
 }
 
 pub mod user {
